define get_regnums as a function and clean code

+++ egr_parser.py
@@ -1,4 +1,3 @@
-import codecs
 import time
 from json import JSONDecodeError
 import concurrent.futures
@@ -6,11 +5,35 @@ import requests
 import json
 import os
 
+PATH = os.path.dirname(os.path.abspath(__file__))
+ENCODING = 'utf-8'
+
+
+def get_regnums():
+    if os.path.isfile(f'{os.path.join(PATH, "ip.json")}') and os.path.isfile(f'{os.path.join(PATH, "jur.json")}'):
+        with open(f'{os.path.join(PATH, "ip.json")}') as f:
+            ip_regnums = [regnum['NM'] for regnum in json.load(f) if regnum['NM'] is not None]
+        with open(f'{os.path.join(PATH, "jur.json")}') as f:
+            jur_regnums = [regnum['NM'] for regnum in json.load(f) if regnum['NM'] is not None]
+    else:
+        print('Download registration numbers.')
+        ip = requests.get("http://egr.gov.by/egrn/API.jsp?TP=1MASK=01000000000000000").json()
+        ip_regnums = [regnum['NM'] for regnum in ip if regnum['NM'] is not None]
+        with open(f'{os.path.join(PATH, "ip.json")}', 'w', encoding=ENCODING) as f:
+            json.dump(ip, f, ensure_ascii=False)
+        print('ip regnums downloaded')
+        jur = requests.get("http://egr.gov.by/egrn/API.jsp?TP=2&MASK=01000000000000000").json()
+        jur_regnums = [regnum['NM'] for regnum in jur if regnum['NM'] is not None]
+        with open(f'{os.path.join(PATH, "jur.json")}', 'w', encoding=ENCODING) as f:
+            json.dump(jur, f, ensure_ascii=False)
+        print('jur regnums downloaded')
+    return ip_regnums, jur_regnums
+
 
 class EgrParser:
-    def __init__(self):
-        self.ip_regnums = []
-        self.jur_regnums = []
+    def __init__(self, ip_regnums, jur_regnums):
+        self.ip_regnums = list(ip_regnums)
+        self.jur_regnums = list(jur_regnums)
         self.all_urls = []
         self.main_list = {}
         self.common_methods = ['getAllAddressByRegNum', 'getAllVEDByRegNum', 'getBaseInfoByRegNum',
@@ -22,6 +45,10 @@ class EgrParser:
     def load_url(url):
         resp = requests.get(url)
         return resp, url
+
+    def server_check(self):
+        resp, _ = self.load_url('http://egr.gov.by/api/v2/egr/getAllAddressByRegNum/100059271')
+        resp.raise_for_status()
 
     def get_jsons(self, urls):
         with concurrent.futures.ThreadPoolExecutor(max_workers=100) as executor:
@@ -37,47 +64,27 @@ class EgrParser:
                     self.all_urls.remove(url)
                 except JSONDecodeError:
                     self.all_urls.remove(url)
-                except Exception as e:
-                    pass
-                    # print(f"Unable to get url due to {e.__class__}.")
+                # except Exception as e:
+                #     print(f"Unable to get url due to {e.__class__}.")
 
-    def get_regnums(self):
-        try:
-            with open(f'{os.path.dirname(os.path.abspath(__file__))}\\ip.json') as f:
-                self.ip_regnums = [regnum['NM'] for regnum in json.load(f)]
-            with open(f'{os.path.dirname(os.path.abspath(__file__))}\\jur.json') as f:
-                self.jur_regnums = [regnum['NM'] for regnum in json.load(f)]
-        except IOError:
-            print('Download registration numbers.')
-            ip = requests.get("http://egr.gov.by/egrn/API.jsp?TP=1MASK=01000000000000000").json()
-            self.ip_regnums = [regnum['NM'] for regnum in ip]
-            with open(f'{os.path.dirname(os.path.abspath(__file__))}\\ip.json', 'w', encoding="utf-8") as f:
-                json.dump(ip, f, ensure_ascii=False)
-            print('ip regnums downloaded')
-            jur = requests.get("http://egr.gov.by/egrn/API.jsp?TP=2&MASK=01000000000000000").json()
-            self.jur_regnums = [regnum['NM'] for regnum in jur]
-            with open(f'{os.path.dirname(os.path.abspath(__file__))}\\jur.json', 'w', encoding="utf-8") as f:
-                json.dump(jur, f, ensure_ascii=False)
+    def create_urls(self, regnums, method):
+        for regnum in regnums:
+            for method in self.common_methods + method:
+                self.all_urls.append(f'http://egr.gov.by/api/v2/egr/{method}/{regnum}')
 
     def get_urls(self):
-        self.get_regnums()
-        try:
-            with open(f'{os.path.dirname(os.path.abspath(__file__))}\\urls.json', 'r', encoding='utf-8') as f:
+        if os.path.isfile(f'{os.path.join(PATH, "urls.json")}'):
+            with open(f'{os.path.join(PATH, "urls.json")}', 'r', encoding=ENCODING) as f:
                 self.all_urls = json.load(f)
-        except IOError:
+        else:
             print('Create urls.')
-            for regnum in self.ip_regnums:
-                if regnum is not None:
-                    for method in self.common_methods + self.ip_methods:
-                        self.all_urls.append(f'http://egr.gov.by/api/v2/egr/{method}/{regnum}')
-            for regnum in self.jur_regnums:
-                if regnum is not None:
-                    for method in self.common_methods + self.jur_methods:
-                        self.all_urls.append(f'http://egr.gov.by/api/v2/egr/{method}/{regnum}')
-            with open(f'{os.path.dirname(os.path.abspath(__file__))}\\urls.json', 'w', encoding="utf-8") as f:
+            self.create_urls(self.ip_regnums, self.ip_methods)
+            self.create_urls(self.jur_regnums, self.jur_methods)
+            with open(f'{os.path.join(PATH, "urls.json")}', 'w', encoding=ENCODING) as f:
                 json.dump(self.all_urls, f, ensure_ascii=False)
 
     def get_data(self):
+        self.server_check()
         self.get_urls()
         time1 = time.time()
         while len(self.all_urls) > 0:
@@ -92,12 +99,13 @@ class EgrParser:
 
     def save_data(self):
         for i in range(0, len(self.main_list), 20000):
-            with codecs.open(f'{os.path.dirname(os.path.abspath(__file__))}\\jsons\\data{int(i / 20000 + 1)}.json', 'w', 'utf-8') as f:
-                json.dump(self.main_list[i:i+20000], f, ensure_ascii=False)
+            with open(os.path.join(PATH, 'jsons', f'data{int(i / 20000 + 1)}.json'), 'w', encoding=ENCODING) as f:
+                json.dump(self.main_list[i:i + 20000], f, ensure_ascii=False)
 
 
 def main():
-    egr = EgrParser()
+    ip_regnums, jur_regnums = get_regnums()
+    egr = EgrParser(ip_regnums, jur_regnums)
     time1 = time.time()
     try:
         egr.get_data()
